Generalized metaclass
