+++ const/const_obj.py
@@ -6,18 +6,9 @@
 import sys
 sys.path.append("../")
 
+from pcad_types.singleton import singleton
 from const.const_strat import const_strat_base
-
-class const_obj_meta(type):
-    _instances = {}
-
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            instance = super().__call__(*args, **kwargs)
-            cls._instances[cls] = instance
-        return cls._instances[cls]
-
-class const_obj_base(metaclass=const_obj_meta):
+class const_obj_base(metaclass=singleton):
     def __init__(self) -> None:
         self.strategy:const_strat_base = None
         pass
