+++ test/test_cube.py
@@ -32,3 +32,5 @@ cube.add(pcad.Dimensioning(pcad.Point(10,0,0),pcad.Point(10,0,10),plane="-yz",te
 cube.add(pcad.Dimensioning(pcad.Point(0,0,0),pcad.Point(0,0,10),plane="-yz",text="-yz"))
 
 prog_parse.exam_execute([cube])
+
+
